+++ src/components/index.ts
@@ -27,6 +27,7 @@ export * from "./grid";
 export * from "./hover-card";
 export * from "./input";
 export * from "./label";
+export * from "./logo";
 export * from "./pagination";
 export * from "./popover";
 export * from "./progress";